Punting on the interrupt refactor for now.

+++ src/io/interrupts.rs
@@ -76,8 +76,8 @@ fn cpu_exception_handler(ctx: &InterruptContext) {
 pub fn rust_interrupt_handler(ctx: &InterruptContext) {
 
     ::state().interrupt_count[ctx.int_id as usize] += 1;
-/*
-    match ctx.int_id {
+
+    match ctx.int_id as u8 {
         0x00...0x0F => cpu_exception_handler(ctx),
         TIMER_INTERRUPT => {
             timer::timer_interrupt();
@@ -104,7 +104,7 @@ pub fn rust_interrupt_handler(ctx: &InterruptContext) {
             ::state().interrupt_count[0 as usize] += 1;
         }
     }
-*/
+
     unsafe {
         PICS.lock().notify_end_of_interrupt(ctx.int_id as u8);
     }
